Refactored Feed/History * Deleted Feed Scene * Initial Refactored
