+++ project/app/src/main/java/com/nexte/nexte/FeedScene/FeedActivity.kt
@@ -1,4 +1,4 @@
-package com.nexte.nexte.HistoryScene
+package com.nexte.nexte.FeedScene
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
@@ -8,9 +8,14 @@ import kotlinx.android.synthetic.main.activity_history.*
 /*
 Class responsible to exhibit player history
  */
-class HistoryActivity : AppCompatActivity(), HistoryDisplayLogic {
 
-    var historyInteractor: HistoryInteractor?= null // reference to the interactor
+interface FeedDisplayLogic {
+    fun displayFeed(viewModel: HistoryModel.ViewModel)
+}
+
+class FeedActivity : AppCompatActivity(), FeedDisplayLogic {
+
+    var interactor: FeedInteractor?= null // reference to the interactor
 
     /*
     This is the function called when the activity is created, and it is responsible to initiate the request
@@ -18,30 +23,30 @@ class HistoryActivity : AppCompatActivity(), HistoryDisplayLogic {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_history)
-        setupHistoryScene()
+        setupFeedScene()
 
         val request = HistoryModel.Request("Helena", this)
-
-        historyInteractor?.getPlayerGames(request)
+        interactor?.fetchFeed(request)
     }
 
     /*
     This method is responsible to set up all the references of this scene
      */
-    private fun setupHistoryScene(){
-        val viewScene = this
-        val interactor = HistoryInteractor()
-        val presenter = HistoryPresenter()
+    private fun setupFeedScene() {
 
-        viewScene.historyInteractor = interactor
+        val view = this
+        val interactor = FeedInteractor()
+        val presenter = FeedPresenter()
+
+        view.interactor = interactor
         interactor.presenter = presenter
-        presenter.viewScene = viewScene
+        presenter.viewScene = view
     }
 
     /*
     Method responsible to receive the viewModel from presenter and show it to the user
     */
-    override fun displayPlayerMatches(viewModel: HistoryModel.ViewModel) {
+    override fun displayFeed(viewModel: HistoryModel.ViewModel) {
         historyListView.adapter = viewModel.adapter
     }
 }
