+++ project/app/src/main/java/com/nexte/nexte/FeedScene/FeedInteractor.kt
@@ -1,22 +1,27 @@
 package com.nexte.nexte.FeedScene
 
 /**
- * Created by leticia on 27/03/18.
+ * Created by helena on 03/04/18.
  */
 
 interface FeedBusinessLogic {
-
-    fun recentGames(request: FeedModel.Request)
+    fun fetchFeed(request: HistoryModel.Request)
 }
-
+/*
+This class is responsible for the communication between worker and presenter
+ */
 class FeedInteractor : FeedBusinessLogic {
 
-    var worker = FeedWorker()
-    var presenter : FeedPresentationLogic? = null
+    var worker = FeedWorker() // Reference to worker
+    var presenter: FeedPresentationLogic? = null // Reference to presenter
 
-    override fun recentGames(request: FeedModel.Request) {
-        worker.getGame(request) { response ->
-            this.presenter?.presentLastGame(response)
+    /*
+    This method is responsible to get the response of the worker and send it to the presenter
+     */
+    override fun fetchFeed(request: HistoryModel.Request) {
+
+        worker.requestFeedData(request) { response ->
+            presenter?.formatFeed(response)
         }
     }
 }
