--- project/app/src/main/java/com/nexte/nexte/FeedScene/FeedView.kt
@@ -1,10 +1,0 @@
-package com.nexte.nexte.FeedScene
-
-/**
- * Created by larissa on 27/03/18.
- */
-
-interface FeedDisplayLogic {
-
-    fun displayRecentGames(viewModel: FeedModel.ViewModel)
-}
