--- project/app/src/main/java/com/nexte/nexte/HistoryScene/HistoryView.kt
@@ -1,9 +1,0 @@
-package com.nexte.nexte.HistoryScene
-
-/**
- * Created by helena on 03/04/18.
- */
-
-interface HistoryDisplayLogic {
-    fun displayPlayerMatches(viewModel: HistoryModel.ViewModel)
-}
