+++ project/app/src/main/java/com/nexte/nexte/RankingScene/RankingWorker.kt
@@ -5,7 +5,6 @@ package com.nexte.nexte.RankingScene
  */
 
 class RankingWorker {
-    constructor()
 
     fun getUsersInRanking(request: RankingModel.Request, completion: (RankingModel.Response) -> Unit) {
