+++ project/app/src/main/res/layout/activity_history.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:alpha="1"
-    tools:context="com.nexte.nexte.HistoryScene.HistoryActivity">
+    tools:context="com.nexte.nexte.FeedScene.FeedActivity">
 
     <ListView
         android:id="@+id/historyListView"
